последняя сборка

+++ ewm-service/src/main/java/ru/practicum/explore/service/CategoryServiceImpl.java
@@ -2,13 +2,17 @@ package ru.practicum.explore.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.explore.dto.category.CategoryDto;
 import ru.practicum.explore.dto.category.NewCategoryDto;
+import ru.practicum.explore.error.exception.ConflictException;
 import ru.practicum.explore.error.exception.NotFoundException;
 import ru.practicum.explore.mapper.CategoryMapper;
 import ru.practicum.explore.model.Category;
 import ru.practicum.explore.repository.CategoryRepository;
+import ru.practicum.explore.repository.EventRepository;
 
 import java.util.List;
 
@@ -18,10 +22,14 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Transactional
     @Override
     public CategoryDto createCategory(NewCategoryDto newCategoryDto) {
+        if (categoryRepository.existsByName(newCategoryDto.getName())) {
+            throw new ConflictException("Категория с именем '" + newCategoryDto.getName() + "' уже существует");
+        }
         Category category = categoryRepository.save(Category.builder()
                 .name(newCategoryDto.getName())
                 .build());
@@ -32,6 +40,9 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public CategoryDto updateCategory(CategoryDto categoryDto, Long catId) {
         Category category = categoryRepository.findById(catId).orElseThrow(NotFoundException::new);
+        if (categoryRepository.existsByName(categoryDto.getName()) && (!category.getName().equals(categoryDto.getName()))) {
+            throw new ConflictException("Категория с именем '" + categoryDto.getName() + "' уже существует");
+        }
         category.setName(categoryDto.getName());
         categoryDto.setId(category.getId());
         return categoryDto;
@@ -40,6 +51,10 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional
     @Override
     public void deleteCategory(Long catId) {
+        if (eventRepository.existsByCategoryId(catId)) {
+            throw new ConflictException("Нельзя удалить категорию" + catId + " потому что она привязана " +
+                    "к событию");
+        }
         categoryRepository.findById(catId).orElseThrow(NotFoundException::new);
         categoryRepository.deleteById(catId);
     }
@@ -52,10 +67,8 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public List<CategoryDto> getCategories(Integer from, Integer size) {
-        List<Category> allCategories = categoryRepository.findAll();
-        return allCategories.stream()
-                .skip(from)
-                .limit(size)
+        Pageable pageable = PageRequest.of(from / size, size);
+        return categoryRepository.findAll(pageable).stream()
                 .map(CategoryMapper::categoryDto)
                 .toList();
     }
